DONE with making Fav button with api also renamed  FavsController to ApiFavsController, lots of changed

+++ app/Http/Controllers/ApiFavsController.php
@@ -3,9 +3,10 @@
 namespace App\Http\Controllers;
 
 use App\Models\Fav;
+use App\Models\Recipe;
 use Illuminate\Http\Request;
 
-class FavsController extends Controller
+class ApiFavsController extends Controller
 {
     /**
      * Store a newly created resource in storage.
@@ -15,7 +16,9 @@ class FavsController extends Controller
      */
     public function store($recipe_id)
     {
-        // $request->validate(['recipe_id' => 'required|numeric']);
+        if (!$recipe_id || !is_numeric($recipe_id) || Recipe::whereId($recipe_id)->doesntExist()) {
+            return response('fail');
+        }
 
         if (Fav::where([['user_id', user()->id], ['recipe_id', $recipe_id]])->exists()) {
             Fav::where([['user_id', user()->id], ['recipe_id', $recipe_id]])->delete();
